feat: Enable configuration of multiple cluster secrets (the first is used for encryption if present)
